+++ example/src/@kickstartds/gatsby-theme-kickstartds/components/BlogDetailPage.tsx
@@ -1,22 +1,22 @@
 import React from 'react';
 import { FunctionComponent } from 'react';
 
-import { Divider } from "@kickstartds/base/lib/divider";
-import { Section } from "@kickstartds/base/lib/section";
-import { TeaserBox } from "@kickstartds/base/lib/teaser-box";
-import { Contact } from "@kickstartds/content/lib/contact";
-import { Storytelling } from "@kickstartds/content/lib/storytelling";
+import { Divider } from '@kickstartds/base/lib/divider';
+import { Section } from '@kickstartds/base/lib/section';
+import { TeaserBox } from '@kickstartds/base/lib/teaser-box';
+import { Contact } from '@kickstartds/base/lib/contact';
+import { Storytelling } from '@kickstartds/content/lib/storytelling';
 
 // TODO this (`Html`) should be in use, instead of `dangerouslySetInnerHTML` below
-import { Html } from "@kickstartds/base/lib/html";
-import { PostHead } from "@kickstartds/blog/lib/post-head";
-import { PostAside } from "@kickstartds/blog/lib/post-aside";
-import { PostShareBar } from "@kickstartds/blog/lib/post-share-bar";
+import { Html } from '@kickstartds/base/lib/html';
+import { PostHead } from '@kickstartds/blog/lib/post-head';
+import { PostAside } from '@kickstartds/blog/lib/post-aside';
+import { PostShareBar } from '@kickstartds/blog/lib/post-share-bar';
 
 import { Layout } from './Layout';
 import { ScrollSpy } from './ScrollSpy';
 
-import { useStaticQuery, graphql } from "gatsby";
+import { useStaticQuery, graphql } from 'gatsby';
 
 export const BlogDetailPage: FunctionComponent<any> = ({
   postHead,
@@ -54,27 +54,37 @@ export const BlogDetailPage: FunctionComponent<any> = ({
           width="wide"
           mode="list"
           headline={{
-            level: "p",
-            align: "center",
-            content: "",
-            spaceAfter: "none",
+            level: 'p',
+            align: 'center',
+            content: '',
+            spaceAfter: 'minimum',
           }}
           spaceBefore="small"
-          spaceAfter="default">
-
+          spaceAfter="default"
+        >
           {postAside && <PostAside {...postAside} />}
 
           <div className="c-post__content">
             {postHead && <PostHead {...postHead} />}
             {/* <div>TODO shouldn't need to dangerouslySetInnerHTML here</div> */}
-            {postBody && <div className="c-html c-rich-text c-post-text" dangerouslySetInnerHTML={{__html: postBody.html}} /> }
+            {postBody && (
+              <div
+                className="c-html c-rich-text c-post-text"
+                dangerouslySetInnerHTML={{ __html: postBody.html }}
+              />
+            )}
             {postShareBar && <PostShareBar {...postShareBar} />}
           </div>
         </Section>
 
-        <Section width="wide" spaceBefore="none" spaceAfter="none" align="center">
+        <Section
+          width="wide"
+          spaceBefore="none"
+          spaceAfter="none"
+          align="center"
+        >
           <Divider />
-        </Section> 
+        </Section>
 
         <Section
           width="narrow"
@@ -83,10 +93,10 @@ export const BlogDetailPage: FunctionComponent<any> = ({
           spaceAfter="small"
           gutter="large"
           headline={{
-            level: "h2",
-            spaceAfter: "none",
-            content: "Get to know us",
-            align: "center",
+            level: 'h2',
+            spaceAfter: 'minimum',
+            content: 'Get to know us',
+            align: 'center',
           }}
         >
           <Storytelling
@@ -99,22 +109,22 @@ export const BlogDetailPage: FunctionComponent<any> = ({
             }}
             box={{
               headline: {
-                content: "Find out if kickstartDS suits your project",
-                level: "h2",
-                spaceAfter: "none",
-                align: "left",
+                content: 'Find out if kickstartDS suits your project',
+                level: 'h2',
+                spaceAfter: 'minimum',
+                align: 'left',
               },
-              text: "Get in contact with our team",
+              text: 'Get in contact with our team',
               link: {
-                label: "Talk to us now",
-                href: "https://www.kickstartDS.com",
-                variant: "solid",
-                size: "medium",
+                label: 'Talk to us now',
+                href: 'https://www.kickstartDS.com',
+                variant: 'solid',
+                size: 'medium',
                 iconAfter: true,
                 icon: {
-                  icon: "chevron-right",
+                  icon: 'chevron-right',
                 },
-              }
+              },
             }}
           />
         </Section>
@@ -125,11 +135,12 @@ export const BlogDetailPage: FunctionComponent<any> = ({
           width="narrow"
           background="default"
           headline={{
-            level: "p",
-            align: "center",
-            content: "",
-            spaceAfter: "none",
-          }}>
+            level: 'p',
+            align: 'center',
+            content: '',
+            spaceAfter: 'minimum',
+          }}
+        >
           <Contact {...postContact} />
         </Section>
 
@@ -140,21 +151,22 @@ export const BlogDetailPage: FunctionComponent<any> = ({
           width="wide"
           background="accent"
           headline={{
-            level: "p",
-            align: "left",
-            styleAs: "h2",
-            content: "Dig deeper ⛏️",
-            spaceAfter: "none",
+            level: 'p',
+            align: 'left',
+            styleAs: 'h2',
+            content: 'Dig deeper ⛏️',
+            spaceAfter: 'minimum',
           }}
-          spaceAfter="none">
+          spaceAfter="none"
+        >
           <TeaserBox
             ratio="16:9"
             image={images.blog.publicURL}
             link={{
               size: 'small',
               href: '/blog/',
-              label: "Read more",
-              variant: "outline",
+              label: 'Read more',
+              variant: 'outline',
             }}
             imageSpacing
             text="Have a look through our other blog posts about all aspects kickstartDS, technical or not."
@@ -166,8 +178,8 @@ export const BlogDetailPage: FunctionComponent<any> = ({
             link={{
               size: 'small',
               href: '/storybook/',
-              label: "Explore components",
-              variant: "outline",
+              label: 'Explore components',
+              variant: 'outline',
             }}
             imageSpacing
             text="Discover all the different components included with kickstartDS, especially their controls and tokens."
@@ -179,8 +191,8 @@ export const BlogDetailPage: FunctionComponent<any> = ({
             link={{
               size: 'small',
               href: '/',
-              label: "Learn more",
-              variant: "outline",
+              label: 'Learn more',
+              variant: 'outline',
             }}
             imageSpacing
             text="Explore kickstartDS, and how it can help your team create consistent interfaces super fast."