+++ gatsby-transformer-kickstartds-contentful/CHANGELOG.md
@@ -1,3 +1,26 @@
+# v2.4.0 (Sat Apr 22 2023)
+
+#### 🚀 Enhancement
+
+- Update glossary, appearances, and showcases pages [#105](https://github.com/kickstartDS/gatsby-theme-kickstartDS/pull/105) ([@julrich](https://github.com/julrich))
+
+#### 🐛 Bug Fix
+
+- fix: use fixed schema tooling, re-fixing headlines in sections ([@julrich](https://github.com/julrich))
+- fix: empty appearances objects ([@julrich](https://github.com/julrich))
+- fix: more generalized header and footer uses ([@julrich](https://github.com/julrich))
+- fix: align showcases, appearances and glossary entries ([@julrich](https://github.com/julrich))
+- fix: get showcases fixed up ([@julrich](https://github.com/julrich))
+- feature: first go at showcase integration ([@julrich](https://github.com/julrich))
+- feature: get appearances running again ([@julrich](https://github.com/julrich))
+- feature: update showcase and appearance pages ([@julrich](https://github.com/julrich))
+
+#### Authors: 1
+
+- Jonas Ulrich ([@julrich](https://github.com/julrich))
+
+---
+
 # v2.3.0 (Sun Nov 27 2022)
 
 #### 🚀 Enhancement